@@ -41,7 +41,11 @@
   "browserslist": "> 0.25%, not dead",
   "dependencies": {
     "@mitodl/course-search-utils": "^2.3.0",
+    "@mitodl/open-api-axios": "^2024.11.14",
+    "@mitodl/smoot-design": "^3.1.0",
+    "@remixicon/react": "^4.6.0",
     "@sentry/browser": "^7.0.0",
+    "@tanstack/react-query": "^5.66.0",
     "array-flat-polyfill": "^1.0.1",
     "bootstrap": "^4.3.1",
     "casual": "^1.6.2",
@@ -60,12 +64,13 @@
     "popper.js": "^1.16.1",
     "promise-polyfill": "^8.3.0",
     "ramda": "^0.30.0",
-    "react": "^16.13.1",
+    "react": "^19.0.0",
     "react-bootstrap": "^1.0.0-beta.15",
     "react-content-loader": "^5.1.2",
-    "react-dom": "^16.13.1",
+    "react-dom": "^19.0.0",
     "react-dotdotdot": "1.3.1",
     "react-infinite-scroller": "^1.2.4",
+    "remixicon": "^4.6.0",
     "sass": "1.77.8",
     "video.js": "^7.11.8",
     "videojs-transcript-ac": "^0.8.8",
@@ -86,7 +91,8 @@
     "@types/lodash.uppercase": "^4.3.6",
     "@types/promise-polyfill": "^6.0.4",
     "@types/ramda": "^0.27.62",
-    "@types/react-dom": "^17.0.11",
+    "@types/react": "^19",
+    "@types/react-dom": "^19",
     "@types/react-infinite-scroller": "^1.2.2",
     "@types/serve-handler": "^6.1.1",
     "@types/sinon": "^10.0.6",
